requirements.txt
faiss-cpu 
pypdf2 
google-generativeai
sentence-transformers 
torch 
numpy 
tqdm
python-dotenv